requirements.txt
torch==1.7.1
torch_geometric==1.6.1
torch-sparse==0.6.8
torch-scatter==2.0.6
numpy==1.19.1
scipy==1.5.2
networkx==2.5.1
matplotlib==3.3.1
dill==0.3.4
python_igraph==0.9.7

